requirements.txt
fastapi
uvicorn
requests
python-dotenv
streamlit
PyMuPDF
